Removed unnecessary code and files
