--- Server/.eslintignore
@@ -1,4 +1,0 @@
-docs/
-public/
-dist/
-node_modules/
