--- Server/.prettierrc
@@ -1,6 +1,0 @@
-{
-  "printWidth": 120,
-  "singleQuote": true,
-  "trailingComma": "none",
-  "endOfLine": "auto"
-}
