+++ Server/package.json
@@ -3,26 +3,18 @@
   "version": "0.0.1",
   "private": true,
   "description": "This is the server side that supports the ACS sample applications",
+  "type": "module",
   "scripts": {
     "build": "npm run clean && webpack --mode=production --env production",
     "clean": "rimraf dist",
-    "start": "ts-node ./bin/www",
+    "start": "node ./src/www",
     "start:dev": "nodemon ./server.js localhost 8080",
-    "start:debug": "node --inspect=9229 -r ts-node/register ./bin/www",
-    "test": "jest",
-    "test:stable": "echo skip",
-    "prettier": "prettier --no-error-on-unmatched-pattern --write --config ./.prettierrc --ignore-path=./.prettierignore \"**/*.js\" \"**/*.jsx\" \"**/*.ts\" \"**/*.tsx\"",
-    "prettier:check": "prettier --no-error-on-unmatched-pattern --check --config ./.prettierrc --ignore-path=./.prettierignore \"**/*.js\" \"**/*.jsx\" \"**/*.ts\" \"**/*.tsx\"",
-    "lint": "eslint */**/*.{ts,tsx}",
-    "lint:fix": "eslint */**/*.{ts,tsx} --fix",
-    "lint:quiet": "eslint */**/*.{ts,tsx} --quiet"
+    "start:debug": "node --inspect=9229 -r node/register ./src/www"
   },
   "dependencies": {
     "@azure/communication-common": "^2.1.0",
     "@azure/communication-identity": "^1.1.0",
     "@azure/core-http": "^2.2.5",
-    "@types/cors": "^2.8.8",
-    "cookie-parser": "~1.4.4",
     "cors": "^2.8.5",
     "debug": "^4.3.4",
     "eslint-plugin-header": "^3.1.0",
@@ -31,36 +23,10 @@
     "morgan": "^1.10.0"
   },
   "devDependencies": {
-    "@types/cookie-parser": "^1.4.3",
-    "@types/copy-webpack-plugin": "^10.1.0",
-    "@types/express": "^4.17.8",
-    "@types/http-errors": "^1.8.2",
-    "@types/jest": "^28.1.6",
-    "@types/morgan": "^1.9.2",
-    "@types/node": "^18.6.3",
-    "@types/supertest": "^2.0.12",
-    "@types/webpack-node-externals": "^2.5.0",
-    "@typescript-eslint/eslint-plugin": "^5.32.0",
-    "@typescript-eslint/parser": "^5.32.0",
     "copy-webpack-plugin": "^11.0.0",
-    "eslint": "^8.21.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-flowtype": "^8.0.3",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "husky": "^8.0.1",
-    "jest": "^28.1.3",
-    "jest-junit": "^14.0.0",
     "nodemon": "^2.0.19",
-    "prettier": "^2.7.1",
-    "pretty-quick": "^3.1.0",
     "rimraf": "^3.0.2",
     "supertest": "^6.2.4",
-    "ts-jest": "^28.0.7",
-    "ts-loader": "^9.3.1",
-    "ts-node": "^10.9.1",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^4.7.4",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0",
     "webpack-node-externals": "^3.0.0"
