+++ Server/webpack.config.js
@@ -1,35 +1,28 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-
 import path from 'path';
 import CopyWebpackPlugin from 'copy-webpack-plugin';
 
+const webConfig = path.resolve('web.config');
+const distDir = path.resolve('dist');
+
 const config = {
   name: 'server',
-  entry: './bin/www.ts',
+  entry: './src/www.js',
   target: 'node',
   output: {
-    path: path.join(__dirname, '/dist'),
+    path: distDir,
     filename: 'server.js'
   },
-  module: {
-    rules: [
-      {
-        test: /\.tsx?$/,
-        use: 'ts-loader',
-        exclude: /node_modules/
-      }
-    ]
-  },
   resolve: {
-    extensions: ['.tsx', '.ts', '.js']
+    extensions: ['.js']
   },
   optimization: {
     minimize: false
   },
   plugins: [
     new CopyWebpackPlugin({
-      patterns: [{ from: path.join(__dirname, 'web.config'), to: path.join(__dirname, 'dist') }]
+      patterns: [{ from: webConfig, to: distDir }]
     })
   ]
 };
